Swapped out old for new doobie mapping
