+++ build.sbt
@@ -35,9 +35,8 @@ lazy val noPublishSettings = Seq(
 lazy val modules: List[ProjectReference] = List(
   core,
   circe,
-  doobie,
   sql,
-  doobie_2,
+  doobie,
   skunk,
   generic,
   demo
@@ -89,28 +88,6 @@ lazy val circe = project
     )
   )
 
-lazy val doobie = project
-  .in(file("modules/doobie"))
-  .enablePlugins(AutomateHeaderPlugin)
-  .disablePlugins(RevolverPlugin)
-  .dependsOn(core, circe)
-  .settings(commonSettings)
-  .settings(
-    name := "gsp-graphql-doobie",
-    fork in Test := true,
-    parallelExecution in Test := false,
-    libraryDependencies ++= Seq(
-      "org.typelevel"     %% "cats-effect"            % catsEffectVersion,
-      "io.chrisdavenport" %% "log4cats-slf4j"         % log4catsVersion,
-      "org.tpolecat"      %% "doobie-core"            % doobieVersion,
-      "org.tpolecat"      %% "doobie-postgres"        % doobieVersion,
-
-      "ch.qos.logback"    %  "logback-classic"        % logbackVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-scalatest"  % testContainersVersion % "test",
-      "com.dimafeng"      %% "testcontainers-scala-postgresql" % testContainersVersion % "test",
-    )
-  )
-
 lazy val sql = project
   .in(file("modules/sql"))
   .enablePlugins(AutomateHeaderPlugin)
@@ -128,14 +105,14 @@ lazy val sql = project
     )
   )
 
-lazy val doobie_2 = project
-  .in(file("modules/doobie-2"))
+lazy val doobie = project
+  .in(file("modules/doobie"))
   .enablePlugins(AutomateHeaderPlugin)
   .disablePlugins(RevolverPlugin)
   .dependsOn(sql % "test->test;compile->compile", circe)
   .settings(commonSettings)
   .settings(
-    name := "gsp-graphql-doobie-2",
+    name := "gsp-graphql-doobie",
     fork in Test := true,
     parallelExecution in Test := false,
     libraryDependencies ++= Seq(
