+++ modules/doobie/src/main/scala/DoobieMappingCompanion.scala
