+++ modules/doobie/src/main/scala/DoobieMonitor.scala
