+++ modules/doobie/src/test/scala/composed/ComposedWorldSpec.scala
@@ -3,139 +3,9 @@
 
 package composed
 
-import io.circe.literal.JsonStringContext
-
 import utils.DatabaseSuite
+import grackle.test.SqlComposedWorldSpec
 
-final class ComposedWorldSpec extends DatabaseSuite {
+final class ComposedWorldSpec extends DatabaseSuite with SqlComposedWorldSpec {
   lazy val mapping = ComposedMapping.fromTransactor(xa)
-
-  test("simple composed query") {
-    val query = """
-      query {
-        country(code: "GBR") {
-          name
-          currencies {
-            code
-            exchangeRate
-          }
-        }
-      }
-    """
-
-    val expected = json"""
-      {
-        "data" : {
-          "country" : {
-            "name" : "United Kingdom",
-            "currencies": [
-              {
-                "code": "GBP",
-                "exchangeRate": 1.25
-              }
-            ]
-          }
-        }
-      }
-    """
-
-    val res = mapping.compileAndRun(query).unsafeRunSync
-    //println(res)
-
-    assert(res == expected)
-  }
-
-  test("simple multiple nested query") {
-    val query = """
-      query {
-        cities(namePattern: "Ame%") {
-          name
-          country {
-            name
-            currencies {
-              code
-              exchangeRate
-            }
-          }
-        }
-      }
-    """
-
-    val expected = json"""
-      {
-        "data": {
-          "cities" : [
-            {
-              "name" : "Amersfoort",
-              "country" : {
-                "name" : "Netherlands",
-                "currencies" : [
-                  {
-                    "code" : "EUR",
-                    "exchangeRate" : 1.12
-                  }
-                ]
-              }
-            },
-            {
-              "name" : "Americana",
-              "country" : {
-                "name" : "Brazil",
-                "currencies" : [
-                  {
-                    "code" : "BRL",
-                    "exchangeRate" : 0.25
-                  }
-                ]
-              }
-            }
-          ]
-        }
-      }
-    """
-
-    val res = mapping.compileAndRun(query).unsafeRunSync
-    //println(res)
-
-    assert(res == expected)
-  }
-
-  test("composed query with introspection") {
-    val query = """
-      query {
-        country(code: "GBR") {
-          __typename
-          name
-          currencies {
-            __typename
-            code
-            exchangeRate
-          }
-        }
-      }
-    """
-
-    val expected = json"""
-      {
-        "data" : {
-          "country" : {
-            "__typename" : "Country",
-            "name" : "United Kingdom",
-            "currencies" : [
-              {
-                "__typename" : "Currency",
-                "code" : "GBP",
-                "exchangeRate" : 1.25
-              }
-            ]
-          }
-        }
-      }
-    """
-
-    val res = mapping.compileAndRun(query).unsafeRunSync
-    //println(res)
-
-    assert(res == expected)
-  }
 }
