+++ modules/doobie/src/test/scala/projection/ProjectionSpec.scala
@@ -3,11 +3,12 @@
 
 package projection
 
-import io.circe.literal.JsonStringContext
+import io.circe.literal._
 
 import utils.DatabaseSuite
+import grackle.test.SqlProjectionSpec
 
-final class ProjectionSpec extends DatabaseSuite {
+final class ProjectionSpec extends DatabaseSuite with SqlProjectionSpec {
   lazy val mapping = ProjectionMapping.fromTransactor(xa)
 
   test("simple projected queries") {
@@ -53,68 +54,6 @@ final class ProjectionSpec extends DatabaseSuite {
     assert(res == expected)
   }
 
-  ignore("projected query with nested query") {
-    val query = """
-      query {
-        level0(filter: { attr: true }) {
-          id
-          level1 {
-            id
-            level2 {
-              id
-              attr
-            }
-          }
-        }
-      }
-    """
-
-    val expected = json"""
-      {
-        "data" : {
-          "level0" : [
-            {
-              "id" : "01",
-              "level1" : [
-                {
-                  "id" : "12",
-                  "level2" : [
-                    {
-                      "id" : "24",
-                      "attr" : true
-                    },
-                    {
-                      "id" : "25",
-                      "attr" : false
-                    }
-                  ]
-                },
-                {
-                  "id" : "13",
-                  "level2" : [
-                    {
-                      "id" : "26",
-                      "attr" : false
-                    },
-                    {
-                      "id" : "27",
-                      "attr" : false
-                    }
-                  ]
-                }
-              ]
-            }
-          ]
-        }
-      }
-    """
-
-    val res = mapping.compileAndRun(query).unsafeRunSync
-    //println(res)
-
-    assert(res == expected)
-  }
-
   test("projected query with nested filtered queries (0)") {
     val query = """
       query {
@@ -145,56 +84,6 @@ final class ProjectionSpec extends DatabaseSuite {
                       "id" : "24",
                       "attr" : true
                     }
-                  ]
-                }
-              ]
-            }
-          ]
-        }
-      }
-    """
-
-    val res = mapping.compileAndRun(query).unsafeRunSync
-    //println(res)
-
-    assert(res == expected)
-  }
-
-  ignore("projected query with nested filtered queries (1)") {
-    val query = """
-      query {
-        level0(filter: { attr: true }) {
-          id
-          level1 {
-            id
-            level2(filter: { attr: true }) {
-              id
-              attr
-            }
-          }
-        }
-      }
-    """
-
-    val expected = json"""
-      {
-        "data" : {
-          "level0" : [
-            {
-              "id" : "01",
-              "level1" : [
-                {
-                  "id" : "12",
-                  "level2" : [
-                    {
-                      "id" : "24",
-                      "attr" : true
-                    }
-                  ]
-                },
-                {
-                  "id" : "13",
-                  "level2" : [
                   ]
                 }
               ]
